+++ lib/generators/authtrail/install_generator.rb
@@ -1,35 +1,51 @@
-# taken from https://github.com/collectiveidea/audited/blob/master/lib/generators/audited/install_generator.rb
-require "rails/generators"
-require "rails/generators/migration"
-require "active_record"
 require "rails/generators/active_record"
 
 module Authtrail
   module Generators
     class InstallGenerator < Rails::Generators::Base
-      include Rails::Generators::Migration
-      source_root File.expand_path("../templates", __FILE__)
+      include ActiveRecord::Generators::Migration
+      source_root File.join(__dir__, "templates")
 
-      # Implement the required interface for Rails::Generators::Migration.
-      def self.next_migration_number(dirname) #:nodoc:
-        next_migration_number = current_migration_number(dirname) + 1
-        if ::ActiveRecord::Base.timestamped_migrations
-          [Time.now.utc.strftime("%Y%m%d%H%M%S"), "%.14d" % next_migration_number].max
-        else
-          "%.3d" % next_migration_number
-        end
-      end
+      class_option :lockbox, type: :boolean
 
       def copy_migration
         migration_template "login_activities_migration.rb", "db/migrate/create_login_activities.rb", migration_version: migration_version
       end
 
+      def copy_templates
+        template "initializer.rb", "config/initializers/authtrail.rb"
+      end
+
       def generate_model
-        template "login_activity_model.rb", "app/models/login_activity.rb"
+        if lockbox?
+          template "model_lockbox.rb", "app/models/login_activity.rb"
+        else
+          template "model.rb", "app/models/login_activity.rb"
+        end
       end
 
       def migration_version
         "[#{ActiveRecord::VERSION::MAJOR}.#{ActiveRecord::VERSION::MINOR}]"
+      end
+
+      def identity_column
+        if lockbox?
+          "t.text :identity_ciphertext\n      t.string :identity_bidx, index: true"
+        else
+          "t.string :identity, index: true"
+        end
+      end
+
+      def ip_column
+        if lockbox?
+          "t.text :ip_ciphertext\n      t.string :ip_bidx, index: true"
+        else
+          "t.string :ip, index: true"
+        end
+      end
+
+      def lockbox?
+        options[:lockbox]
       end
     end
   end